backend/requirements.txt
python-dotenv
requests
google-generativeai
fastapi
uvicorn